added feature struct, expanded language struct

+++ src/data/items.rs
@@ -1,10 +1,10 @@
 #[derive(Default, Debug, Clone)]
 pub struct Item {
-   pub id: Option<u64>,
+   pub id: Option<i64>,
    pub name: String,
    pub class: String,
    pub cost: u16,
    pub damage: Option<(u8, u8)>,
-   pub weight: u16,
+   pub weight: u64,
    pub properties: String
 }
